+++ src/helpers.ts
@@ -1,8 +1,23 @@
+import * as dayjs from 'dayjs';
+
 var register = function (Handlebars) {
     var helpers = {
         gt: function (num1, num2) {
             return num1 > num2
         },
+
+        gte: function(num1, num2){
+            return num1 >= num2
+        },
+        
+        eq: function(arg1, arg2){
+            return arg1 == arg2
+        },
+
+        ifNotEquals: function(arg1, arg2, options) {
+            return (arg1 != arg2) ? options.fn(this) : options.inverse(this);
+        },
+
         capFirstChar: function (string: string) {
             let splitted = string.split(' ')
             for (let i = 0; i < splitted.length; i++) {
@@ -10,6 +25,20 @@ var register = function (Handlebars) {
             }
             return splitted.join(' ')
         },
+
+        formatDate: function(utcFormat: Date, format: string){
+            return dayjs(utcFormat).format(format)
+        },
+
+        fromNow: function(utc){
+            let now = dayjs()
+            let date = dayjs(utc)
+            if(now.diff(date, 'second') < 60) return `${now.diff(date, 'second')} seconds ago`
+            if(now.diff(date, 'minute') < 60) return `${now.diff(date, 'minute')} minutes ago`
+            if(now.diff(date, 'hour') < 24) return `${now.diff(date, 'hour')} hours ago`
+            if(now.diff(date, 'day') < 31) return `${now.diff(date, 'day')} days ago`
+        },
+
         mapGender: function (gender, options) {
             if (gender === "M") {
                 return `<option value="M" selected>Male</option>`
@@ -19,6 +48,7 @@ var register = function (Handlebars) {
             }
             return "yay"
         },
+
         select: function (value, options) {
             return options.fn(this)
                 .split('\n')
@@ -27,8 +57,12 @@ var register = function (Handlebars) {
                     return !RegExp(t).test(v) ? v : v.replace(t, t + ' selected="selected"')
                 })
                 .join('\n')
-        }
+        },
+
+
     };
+
+
 
     if (Handlebars && typeof Handlebars.registerHelper === "function") {
         for (var prop in helpers) {
